# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MathPlatform API",
        "description": "API for MathPlatform",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/students": {
            "get": {
                "tags": [
                    "Admin Students"
                ],
                "summary": "Get all students",
                "description": "Returns a paginated list of all students.",
                "operationId": "getAllStudents",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Students retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Admin access required"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Students"
                ],
                "summary": "Create student manually by admin",
                "description": "Creates a student, enrollment, active subscription, and an approved subscription code. No approval is required when the student is created by an admin.",
                "operationId": "createStudentByAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone",
                                    "parent_phone",
                                    "grade_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmed"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "01012345678"
                                    },
                                    "parent_phone": {
                                        "type": "string",
                                        "example": "01112345678"
                                    },
                                    "grade_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-09-18",
                                        "nullable": true
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2027-09-18",
                                        "nullable": true
                                    },
                                    "code_expires_at": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2027-09-18",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Admin access required"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/students/{student}": {
            "get": {
                "tags": [
                    "Admin Students"
                ],
                "summary": "Get student details",
                "description": "Returns the details of a specific student.",
                "operationId": "getStudentByAdmin",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Admin access required"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin Students"
                ],
                "summary": "Update student",
                "description": "Updates the basic information of a student.",
                "operationId": "updateStudentByAdmin",
                "parameters": [
                    {
                        "name": "student",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "phone",
                                    "parent_phone",
                                    "grade_id"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Ahmed"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Ali"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "01012345678"
                                    },
                                    "parent_phone": {
                                        "type": "string",
                                        "example": "01112345678"
                                    },
                                    "grade_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Admin access required"
                    },
                    "404": {
                        "description": "Student not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter the Sanctum token returned by the admin login.",
                "bearerFormat": "Sanctum Token",
                "scheme": "bearer"
            }
        }
    },
...
}